market_dashboard/requirements.txt
streamlit
pandas
pyyaml            # ל־config
# כל ספריית HTTP/Scraping שתצטרך (requests, feedparser וכו’)
